iot-device-manager/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0
requests==2.31.0
Pillow==10.1.0
paramiko==3.3.1
cryptography==41.0.7
PyJWT==2.8.0
django-extensions==3.2.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.29.7
paho-mqtt==1.6.1
influxdb-client==1.38.0
numpy==1.25.2
pandas==2.1.3
django-ratelimit==4.1.0
django-csp==3.7
django-security==0.18.0
bleach==6.1.0 